+++ src/index.ts
@@ -1,4 +1,4 @@
-export * from './clients/gid-client';
-export * from './clients/gid-client-factory';
-export { default } from './clients/gid-client';
+export * from './clients/gid-issuer-client';
+export * from './clients/gid-issuer-client-factory';
+export { default } from './clients/gid-issuer-client';
 export * from './utils/download-file';